remove entities module
