+++ thetawave_lib/src/boss/entities/mod.rs
@@ -1,0 +1,5 @@
+//! Entities for managing bosses
+
+mod boss;
+
+pub use self::boss::spawn_repeater;
