+++ thetawave_lib/src/boss/mod.rs
@@ -1,4 +1,5 @@
 //! `thetawave_lib` boss module
 
 pub mod components;
+pub mod entities;
 pub mod systems;
