--- thetawave_lib/src/entities/mod.rs
@@ -1,88 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-pub mod barriers;
-pub mod boss;
-pub mod spaceship;
-
-pub use self::{
-    barriers::initialize_arena_barriers, boss::spawn_repeater, spaceship::initialize_spaceship,
-};
-
-#[derive(Clone, Serialize, Deserialize, Debug, Hash, PartialEq, Eq)]
-pub enum SpawnableType {
-    Consumable(ConsumableType),
-    Item(ItemType),
-    Effect(EffectType),
-    Mob(MobType),
-}
-
-#[derive(Clone, Serialize, Deserialize, Debug, Hash, PartialEq, Eq)]
-pub enum MobType {
-    Enemy(EnemyType),
-    Ally(AllyType),
-    Neutral(NeutralType),
-}
-
-#[derive(Clone, Serialize, Deserialize, Debug, Hash, PartialEq, Eq)]
-pub enum EnemyType {
-    Pawn,
-    Drone,
-    StraferRight,
-    StraferLeft,
-    MissileLauncher,
-    Missile,
-    RepeaterBody,
-    RepeaterHead,
-    RepeaterLeftShoulder,
-    RepeaterRightShoulder,
-    RepeaterLeftArm,
-    RepeaterRightArm,
-}
-
-#[derive(Clone, Serialize, Deserialize, Debug, Hash, PartialEq, Eq)]
-pub enum AllyType {
-    Hauler,
-}
-
-#[derive(Clone, Serialize, Deserialize, Debug, Hash, PartialEq, Eq)]
-pub enum NeutralType {
-    MoneyAsteroid,
-}
-
-#[derive(Clone, Serialize, Deserialize, Debug, Hash, PartialEq, Eq)]
-pub enum ConsumableType {
-    DefenseWrench,
-    Money1,
-    Money5,
-    HealthWrench,
-    Armor,
-}
-
-#[derive(Clone, Serialize, Deserialize, Debug, Hash, PartialEq, Eq)]
-pub enum ItemType {
-    SteelBarrel,
-    PlasmaBlasts,
-    HazardousReactor,
-    WarpThruster,
-    Tentaclover,
-    DefenseSatellite,
-    DoubleBarrel,
-    YithianPlague,
-    Spice,
-    EnhancedPlating,
-    StructureReinforcement,
-    BlasterSizeEnhancer,
-    FrequencyAugmentor,
-    TractorBeam,
-    BlastRepeller,
-}
-#[derive(Clone, Serialize, Deserialize, Debug, Hash, PartialEq, Eq)]
-pub enum EffectType {
-    AllyBlastExplosion,
-    EnemyBlastExplosion,
-    PoisonBlastExplosion,
-    CriticalBlastExplosion,
-    MobExplosion,
-    Star, //TODO: implement background stars
-    Giblets(MobType),
-}
