+++ thetawave_lib/src/events/events.rs
@@ -1,6 +1,6 @@
 use crate::{
-    entities::{ConsumableType, ItemType},
     misc::components::{AttractData, AttractorCategory},
+    spawnable::resources::{ConsumableType, ItemType},
 };
 use amethyst::{audio::SourceHandle, core::math::Vector2, ecs::prelude::Entity};
 use std::collections::HashMap;
