+++ thetawave_lib/src/lib.rs
@@ -1,7 +1,6 @@
 pub mod audio;
 pub mod boss;
 pub mod constants;
-pub mod entities;
 pub mod events;
 pub mod misc;
 pub mod motion;
