+++ thetawave_lib/src/misc/entities/mod.rs
@@ -1,0 +1,5 @@
+//! Miscellaneous entities
+
+mod barriers;
+
+pub use self::barriers::initialize_arena_barriers;
