+++ thetawave_lib/src/misc/mod.rs
@@ -1,5 +1,6 @@
 //! `thetawave_lib` misc module
 
 pub mod components;
+pub mod entities;
 pub mod resources;
 pub mod systems;
