+++ thetawave_lib/src/motion/systems/motion2d.rs
@@ -1,11 +1,13 @@
 use crate::{
     constants::{ARENA_HEIGHT, ARENA_MIN_Y},
-    entities::{AllyType, EnemyType, MobType, NeutralType},
     events::AttractionEvent,
     misc::components::AttractorCategory,
     motion::components::{Hitbox2DComponent, Motion2DComponent},
     player::components::PlayerComponent,
-    spawnable::components::{BlastComponent, ConsumableComponent, ItemComponent, MobComponent},
+    spawnable::{
+        components::{BlastComponent, ConsumableComponent, ItemComponent, MobComponent},
+        resources::{AllyType, EnemyType, MobType, NeutralType},
+    },
     tools::distance,
     weapons::BlastType,
 };
