+++ thetawave_lib/src/player/entities/mod.rs
@@ -1,0 +1,5 @@
+//! Entities for managing players
+
+mod spaceship;
+
+pub use self::spaceship::initialize_spaceship;
