+++ thetawave_lib/src/player/mod.rs
@@ -1,5 +1,6 @@
 //! `thetawave_lib` player module
 
 pub mod components;
+pub mod entities;
 pub mod resources;
 pub mod systems;
