+++ thetawave_lib/src/spawn/resources/drops.rs
@@ -1,7 +1,7 @@
 use crate::{
-    entities::SpawnableType,
     spawnable::resources::{
         spawn_spawnable, ConsumablesResource, EffectsResource, ItemsResource, MobsResource,
+        SpawnableType,
     },
     tools::weighted_rng,
     visual::resources::SpriteSheetsResource,
